Move types.GrantedScopes to repos.GrantedScopes
The types package shouldn't import code host specific code. Ideally we'd
want to move GrantedScopes to the extsvc package but this it currently
not possible due to cyclic dependencies.
